+++ packages/form-table/src/public-types.ts
@@ -3,7 +3,6 @@ import type { FormTableComponent, TableRow } from './types.public'
 
 export * from './types.public'
 export { defineFormTableColumns } from './defineFormTableColumns'
-export { createFormTableField } from './createFormTableField'
 
 export declare function createFormTable<TRow extends TableRow = TableRow>(): FormTableComponent<TRow>